修复无关联店铺时，应用crash的bug

+++ platform/ios/ContactLoadingView.h
@@ -2,6 +2,6 @@
 
 @interface ContactLoadingView : UIView
 
--(void) showMemberCount:(int)count;
+-(void) showMemberCount:(NSUInteger)count;
 
 @end
